requirements.txt
qrcode[pil]
Pillow
customtkinter
